+++ src/components/NotificationItem/index.tsx
@@ -4,7 +4,7 @@ import IconKebab from "@/assets/svgs/ic_kebab.svg";
 import HighlightText from "@/src/components/HighlightText/HighlightText";
 import STATUS_KEYWORD_HIGHLIGHT from "@/src/constants/myNotifications";
 import PATH_NAMES from "@/src/constants/pathname";
-import type { MyNotificationsProps } from "@/src/types/my-notifications";
+import type { MyNotificationsProps } from "@/src/types/myNotifications";
 import { formatDateToKorean } from "@/src/utils/calendarFormatDate";
 import { getTimeDiffText } from "@/src/utils/date";
 import clsx from "clsx";
